refactor(node): move `Node` stuff into `node.rs`
