+++ Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "halfin"
 version = "0.4.0"
-description = "A {regtest} bitcoin node runner 🏃‍♂️"
+description = "A {regtest} bitcoin node and indexer runner 🏃‍♂️"
 authors = ["Luis Schwab <dev@luisschwab.net"]
 documentation = "https://docs.rs/halfin"
 repository = "https://github.com/luisschwab/halfin"
-keywords = ["bitcoin", "runner", "testing"]
+keywords = ["bitcoin", "electrum", "runner", "testing"]
 license = "MIT OR Apache-2.0"
 readme = "README.md"
 rust-version = "1.85.0" # MSRV
