+++ src/bitcoind/mod.rs
@@ -56,10 +56,10 @@ use crate::CONNECTION_TIMEOUT;
 use crate::DataDir;
 use crate::Error;
 use crate::IPV4_LOCALHOST;
-use crate::NODE_BUILDING_ATTEMPTS;
-use crate::NODE_BUILDING_INTERVAL;
-use crate::Node;
+use crate::SPAWN_ATTEMPTS;
+use crate::SPAWN_INTERVAL;
 use crate::get_available_port;
+use crate::node::Node;
 use crate::pipe_to_tracing;
 
 /// Name of the wallet created (or loaded) inside every [`BitcoinD`] instance.
@@ -126,7 +126,7 @@ pub struct BitcoinDConf<'a> {
     /// How many times to retry spawning `bitcoind` before giving up.
     ///
     /// Each attempt picks fresh random ports, so transient port-collision
-    /// errors are automatically recovered from. Defaults to [`NODE_BUILDING_ATTEMPTS`].
+    /// errors are automatically recovered from. Defaults to [`SPAWN_ATTEMPTS`].
     pub max_retries: u8,
 }
 
@@ -136,7 +136,7 @@ impl Default for BitcoinDConf<'_> {
             args: vec!["-regtest", "-fallbackfee=0.0001", "-blockfilterindex=1"],
             tmpdir: None,
             staticdir: None,
-            max_retries: NODE_BUILDING_ATTEMPTS,
+            max_retries: SPAWN_ATTEMPTS,
         }
     }
 }
@@ -176,9 +176,9 @@ pub struct BitcoinD {
 
 #[rustfmt::skip]
 impl Node for BitcoinD {
-    fn get_name() -> &'static str { "BitcoinD" }
+    fn get_name() -> &'static str { versions::BITCOIND_NAME }
 
-    fn get_bin_name() -> &'static str { "bitcoind_v_31_0" }
+    fn get_bin_name() -> &'static str { versions::BITCOIND_BIN_NAME }
 
     fn get_p2p_socket(&self) -> SocketAddr { self.get_p2p_socket() }
 
@@ -306,7 +306,7 @@ impl BitcoinD {
 
             // Add a small timeout to let `bitcoind` fail
             // and retry in the case of a port collision.
-            sleep(NODE_BUILDING_INTERVAL);
+            sleep(SPAWN_INTERVAL);
 
             // If the process exited immediately, try again with new ports.
             match process.try_wait() {
