+++ src/bitcoind/versions.rs
@@ -1,3 +1,11 @@
-/// Bitcoin Core v31.0
+/// [`BitcoinD`](crate::bitcoind::BitcoinD)'s human-readable name.
+#[allow(unused)]
+pub(crate) const BITCOIND_NAME: &str = "BitcoinD";
+
+/// [`BitcoinD`](crate::bitcoind::BitcoinD)'s binary name.
+#[allow(unused)]
+pub(crate) const BITCOIND_BIN_NAME: &str = "bitcoind_v_31_0";
+
+/// `bitcoind` v31.0
 #[allow(unused)]
 pub(crate) const BITCOIND_VERSION: &str = "31.0";
