+++ src/electrsd/mod.rs
@@ -52,11 +52,11 @@ use crate::BitcoinD;
 use crate::DataDir;
 use crate::Error;
 use crate::IPV4_LOCALHOST;
-use crate::NODE_BUILDING_ATTEMPTS;
-use crate::NODE_BUILDING_INTERVAL;
-use crate::Node;
 use crate::POLL_INTERVAL;
+use crate::SPAWN_ATTEMPTS;
+use crate::SPAWN_INTERVAL;
 use crate::get_available_port;
+use crate::node::Node;
 use crate::pipe_to_tracing;
 
 /// Bundled `electrs` version metadata.
@@ -128,7 +128,7 @@ pub struct ElectrsDConf<'a> {
     /// How many times to retry spawning `electrs` before giving up.
     ///
     /// Each attempt picks fresh random ports, so transient port-collision
-    /// errors are automatically recovered from. Defaults to [`NODE_BUILDING_ATTEMPTS`].
+    /// errors are automatically recovered from. Defaults to [`SPAWN_ATTEMPTS`].
     pub max_retries: u8,
 }
 
@@ -139,7 +139,7 @@ impl Default for ElectrsDConf<'_> {
             network: "regtest",
             tmpdir: None,
             staticdir: None,
-            max_retries: NODE_BUILDING_ATTEMPTS,
+            max_retries: SPAWN_ATTEMPTS,
         }
     }
 }
@@ -177,10 +177,10 @@ pub struct ElectrsD {
 #[rustfmt::skip]
 impl ElectrsD {
     /// [`ElectrsD`]'s human-readable name.
-    pub fn get_name() -> &'static str { "ElectrsD" }
+    pub fn get_name() -> &'static str { versions::ELECTRS_NAME }
 
     /// [`ElectrsD`]'s binary name.
-    pub fn get_bin_name() -> &'static str { "electrs_v_0_11_1" }
+    pub fn get_bin_name() -> &'static str { versions::ELECTRS_BIN_NAME }
 }
 
 impl ElectrsD {
@@ -312,7 +312,7 @@ impl ElectrsD {
 
             // Add a small timeout to let `electrs` fail
             // and retry in the case of a port collision.
-            sleep(NODE_BUILDING_INTERVAL);
+            sleep(SPAWN_INTERVAL);
 
             // If the process exited immediately, try again with new ports.
             match process.try_wait() {
