+++ src/electrsd/versions.rs
@@ -1,3 +1,11 @@
-/// electrs v0.11.1
+/// [`ElectrsD`](crate::electrsd::ElectrsD)'s human-readable name.
+#[allow(unused)]
+pub(crate) const ELECTRS_NAME: &str = "ElectrsD";
+
+/// [`ElectrsD`](crate::electrsd::ElectrsD)'s binary name.
+#[allow(unused)]
+pub(crate) const ELECTRS_BIN_NAME: &str = "electrs_v_0_11_1";
+
+/// `electrs` v0.11.1
 #[allow(unused)]
 pub(crate) const ELECTRS_VERSION: &str = "0.11.1";
