+++ src/utreexod/versions.rs
@@ -1,3 +1,11 @@
-/// Utreexod v0.6.0
+/// [`UtreexoD`](crate::utreexod::UtreexoD)'s human-readable name.
+#[allow(unused)]
+pub(crate) const UTREEXOD_NAME: &str = "UtreexoD";
+
+/// [`UtreexoD`](crate::utreexod::UtreexoD)'s binary name.
+#[allow(unused)]
+pub(crate) const UTREEXOD_BIN_NAME: &str = "utreexod_v_0_6_0";
+
+/// `utreexod` v0.6.0
 #[allow(unused)]
 pub(crate) const UTREEXOD_VERSION: &str = "0.6.0";
